[15.0][IMP] Mejorando reportes de estados financieros.
