+++ l10n_cu_reports/README.rst
@@ -27,6 +27,7 @@ Bug Tracker
 Dependencias de terceros
 =============
 Odoo Mates: https://github.com/odoomates/odooapps
+OCA: https://github.com/OCA/reporting-engine
 
 Authors
 ==========
