+++ l10n_cu_reports/reports/report_financial.xml
@@ -39,5 +39,39 @@
         <field name="report_name">l10n_cu_reports.report_financial_evab</field>
         <field name="report_file">l10n_cu_reports.report_financial_evab</field>
         <field name="paperformat_id" ref="paperformat_horizontal"/>
-    </record>    
+    </record>
+
+    <record id="action_ncc_accounting_xls_reports_report_bs" model="ir.actions.report">
+        <field name="name">Estado de Situación XLSX</field>
+        <field name="model">account.financial.report</field>
+        <field name="report_type">xlsx</field>
+        <field name="report_name">l10n_cu_reports_xlsx.report_financial_xls_bs</field>
+        <field name="report_file">l10n_cu_reports_xlsx.report_financial_xls_bs</field>
+<!--        <field name="attachment_use">True</field>-->
+    </record>
+    <record id="action_ncc_accounting_xls_reports_report_pl" model="ir.actions.report">
+        <field name="name">Estado de Rendimiento Financiero XLSX</field>
+        <field name="model">account.financial.report</field>
+        <field name="report_type">xlsx</field>
+        <field name="report_name">l10n_cu_reports_xlsx.report_financial_xls_pl</field>
+        <field name="report_file">l10n_cu_reports_xlsx.report_financial_xls_pl</field>
+<!--        <field name="attachment_use">True</field>-->
+    </record>
+    <record id="action_ncc_accounting_xls_reports_report_ege" model="ir.actions.report">
+        <field name="name">Estado de Rendimiento Financiero XLSX</field>
+        <field name="model">account.financial.report</field>
+        <field name="report_type">xlsx</field>
+        <field name="report_name">l10n_cu_reports_xlsx.report_financial_xls_ege</field>
+        <field name="report_file">l10n_cu_reports_xlsx.report_financial_xls_ege</field>
+<!--        <field name="attachment_use">True</field>-->
+    </record>
+
+    <record id="action_ncc_accounting_xls_reports_report_evab" model="ir.actions.report">
+        <field name="name">Estado de Valor Agregado Bruto XLSX</field>
+        <field name="model">account.financial.report</field>
+        <field name="report_type">xlsx</field>
+        <field name="report_name">l10n_cu_reports_xlsx.report_financial_xls_evab</field>
+        <field name="report_file">l10n_cu_reports_xlsx.report_financial_xls_evab</field>
+<!--        <field name="attachment_use">True</field>-->
+    </record>
 </odoo>
