+++ l10n_cu_reports/wizards/account_report.py
@@ -8,7 +8,7 @@ class AccountingReport(models.TransientModel):
     _inherit = 'accounting.report'
 
     target_move = fields.Selection(selection=[('posted', 'All Posted Entries'),
-                                    ('all', 'All Entries')])
+                                              ('all', 'All Entries')])
     date_to = fields.Date(required=True, default=lambda self: datetime.now())
     
     display_account = fields.Selection(selection=[("all", "All"), ("not_zero", "With balance is not equal to 0")],
@@ -20,22 +20,55 @@ class AccountingReport(models.TransientModel):
     ], string='Display details', default='no_detail')
 
     def _print_report(self, data):
+        # data['form'].update(self.read(['debit_credit', 'display_account', 'display_detail',
+        #                                'account_report_id', 'target_move'])[0])
+        # context = self._context
+        # if context.get('efe', '') == '5921-04':
+        #     return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_pl').\
+        #     report_action(self, data=data, config=False)
+        # elif context.get('efe', '') == '5920-04':
+        #     return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_bs'). \
+        #     report_action(self, data=data, config=False)
+        # elif context.get('efe', '') == '5924-04':
+        #     return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_ege'). \
+        #     report_action(self, data=data, config=False)
+        # elif context.get('efe', '') == '5926-04':
+        #     return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_evab'). \
+        #     report_action(self, data=data, config=False)
+        # else:
+        #     return False
         data['form'].update(self.read(['debit_credit', 'display_account', 'display_detail',
                                        'account_report_id', 'target_move'])[0])
         context = self._context
-        if context.get('efe', '') == '5921-04':
-            return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_pl').\
-            report_action(self, data=data, config=False)
-        elif context.get('efe', '') == '5920-04':
-            return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_bs'). \
-            report_action(self, data=data, config=False)
+        if context.get('efe', '') == '5920-04':
+            if context.get('type', '') == 'xlsx':
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_xls_reports_report_bs'). \
+                    report_action(self, data=data, config=False)
+            else:
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_bs'). \
+                    report_action(self, data=data, config=False)
+        elif context.get('efe', '') == '5921-04':
+            if context.get('type', '') == 'xlsx':
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_xls_reports_report_pl'). \
+                    report_action(self, data=data, config=False)
+            else:
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_pl'). \
+                    report_action(self, data=data, config=False)
         elif context.get('efe', '') == '5924-04':
-            return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_ege'). \
-            report_action(self, data=data, config=False)
+            if context.get('type', '') == 'xlsx':
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_xls_reports_report_ege'). \
+                    report_action(self, data=data, config=False)
+            else:
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_ege'). \
+                    report_action(self, data=data, config=False)
         elif context.get('efe', '') == '5926-04':
-            return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_evab'). \
-            report_action(self, data=data, config=False)
+            if context.get('type', '') == 'xlsx':
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_xls_reports_report_evab'). \
+                    report_action(self, data=data, config=False)
+            else:
+                return self.env.ref('l10n_cu_reports.action_ncc_accounting_pdf_reports_report_evab'). \
+                    report_action(self, data=data, config=False)
         else:
-            return False
+            return {}
