+++ l10n_cu_reports/wizards/account_report.xml
@@ -13,7 +13,8 @@
                     <field name="account_report_id" invisible="1"/>
                 </group>
                 <footer>
-                    <button name="check_report" string="Print" type="object" default_focus="1" class="oe_highlight"/>
+                    <button name="check_report" string="Print" type="object" context="{'type': 'pdf'}" default_focus="1" class="oe_highlight"/>
+                    <button name="check_report" string="XLSX" type="object" context="{'type': 'xlsx'}" class="oe_highlight"/>
                     <button string="Cancel" class="btn btn-secondary" special="cancel"/>
                 </footer>
             </form>
