--- l10n_cu_reports_xlsx/README.rst
@@ -1,29 +1,0 @@
-================================
-Cuba - Proformas Estados Financieros (XLSX)
-================================
-
-Configuración
-=============
-
-
-
-Uso
-=====
-Reportes de estados financieros legales en XLSX.
-
-
-Bug Tracker
-===========
-
-Dependencias de terceros
-=============
-OCA: https://github.com/OCA/reporting-engine
-
-Authors
-==========
-* Comunidad Cubana de Odoo
-
-Contributors
-=============
-
-* Segu
