--- l10n_cu_reports_xlsx/__init__.py
@@ -1,3 +1,0 @@
-from . import reports
-from . import wizards
-
