--- l10n_cu_reports_xlsx/reports/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import report_financial_xls
