--- l10n_cu_reports_xlsx/wizards/__init__.py
@@ -1,1 +1,0 @@
-from . import account_report
